requirements.txt
asgiref==3.8.1
black==25.1.0
certifi==2025.6.15
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
Django==5.2.1
django-bootstrap5==25.1
djangorestframework==3.14.0
django-filter==23.5
drf-spectacular==0.27.0
et_xmlfile==2.0.0
idna==3.10
isort==6.0.1
mypy_extensions==1.1.0
numpy==2.3.5
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
reportlab==4.4.1
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.5.0
